Create QueueCreator
