+++ scripts/core/index.ts
@@ -9,3 +9,4 @@ export * from './guideManager'
 export * from './guide'
 export * from './apiChannel'
 export * from './apiClient'
+export * from './queueCreator'
